+++ src/CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Improvements and Fixes
  * Added an option to change map size. To do that open a new "Settings" menu, available under the cog button in the top-right corner of the opened map pane;
  * It's now possible to interact with the map window without first having to manually tab in. Specifically for: select instance -> change a var -> select another instance;
+ * Added a preference to adjust the application framerate;
  * Minor GUI improvements;
  * Fixed attempting to pick/delete item pixel shifted off map does not work.